Removed challenge sensors
