+++ srunner/autoagents/agent_wrapper.py
@@ -15,8 +15,7 @@ import os
 
 import carla
 
-from srunner.challenge.envs.scene_layout_sensors import SceneLayoutReader, ObjectFinder
-from srunner.challenge.envs.sensor_interface import CallBack, CANBusSensor, HDMapReader
+from srunner.autoagents.sensor_interface import CallBack
 from srunner.autoagents.autonomous_agent import Track
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
 
@@ -50,51 +49,37 @@ class AgentWrapper(object):
         """
         bp_library = CarlaDataProvider.get_world().get_blueprint_library()
         for sensor_spec in self._agent.sensors():
-            # These are the pseudosensors (not spawned)
-            if sensor_spec['type'].startswith('sensor.scene_layout'):
-                # Static sensor that gives you the entire information from the world (Just runs once)
-                sensor = SceneLayoutReader(CarlaDataProvider.get_world())
-            elif sensor_spec['type'].startswith('sensor.object_finder'):
-                # This sensor returns the position of the dynamic objects in the scene.
-                sensor = ObjectFinder(CarlaDataProvider.get_world(), sensor_spec['reading_frequency'])
-            elif sensor_spec['type'].startswith('sensor.can_bus'):
-                # The speedometer pseudo sensor is created directly here
-                sensor = CANBusSensor(vehicle, sensor_spec['reading_frequency'])
-            elif sensor_spec['type'].startswith('sensor.hd_map'):
-                # The HDMap pseudo sensor is created directly here
-                sensor = HDMapReader(vehicle, sensor_spec['reading_frequency'])
             # These are the sensors spawned on the carla world
-            else:
-                bp = bp_library.find(str(sensor_spec['type']))
-                if sensor_spec['type'].startswith('sensor.camera'):
-                    bp.set_attribute('image_size_x', str(sensor_spec['width']))
-                    bp.set_attribute('image_size_y', str(sensor_spec['height']))
-                    bp.set_attribute('fov', str(sensor_spec['fov']))
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
-                                                     roll=sensor_spec['roll'],
-                                                     yaw=sensor_spec['yaw'])
-                elif sensor_spec['type'].startswith('sensor.lidar'):
-                    bp.set_attribute('range', str(sensor_spec['range']))
-                    bp.set_attribute('rotation_frequency', str(sensor_spec['rotation_frequency']))
-                    bp.set_attribute('channels', str(sensor_spec['channels']))
-                    bp.set_attribute('upper_fov', str(sensor_spec['upper_fov']))
-                    bp.set_attribute('lower_fov', str(sensor_spec['lower_fov']))
-                    bp.set_attribute('points_per_second', str(sensor_spec['points_per_second']))
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
-                                                     roll=sensor_spec['roll'],
-                                                     yaw=sensor_spec['yaw'])
-                elif sensor_spec['type'].startswith('sensor.other.gnss'):
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation()
+            bp = bp_library.find(str(sensor_spec['type']))
+            if sensor_spec['type'].startswith('sensor.camera'):
+                bp.set_attribute('image_size_x', str(sensor_spec['width']))
+                bp.set_attribute('image_size_y', str(sensor_spec['height']))
+                bp.set_attribute('fov', str(sensor_spec['fov']))
+                sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
+                                                    z=sensor_spec['z'])
+                sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
+                                                    roll=sensor_spec['roll'],
+                                                    yaw=sensor_spec['yaw'])
+            elif sensor_spec['type'].startswith('sensor.lidar'):
+                bp.set_attribute('range', str(sensor_spec['range']))
+                bp.set_attribute('rotation_frequency', str(sensor_spec['rotation_frequency']))
+                bp.set_attribute('channels', str(sensor_spec['channels']))
+                bp.set_attribute('upper_fov', str(sensor_spec['upper_fov']))
+                bp.set_attribute('lower_fov', str(sensor_spec['lower_fov']))
+                bp.set_attribute('points_per_second', str(sensor_spec['points_per_second']))
+                sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
+                                                    z=sensor_spec['z'])
+                sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
+                                                    roll=sensor_spec['roll'],
+                                                    yaw=sensor_spec['yaw'])
+            elif sensor_spec['type'].startswith('sensor.other.gnss'):
+                sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
+                                                    z=sensor_spec['z'])
+                sensor_rotation = carla.Rotation()
 
-                # create sensor
-                sensor_transform = carla.Transform(sensor_location, sensor_rotation)
-                sensor = CarlaDataProvider.get_world().spawn_actor(bp, sensor_transform, vehicle)
+            # create sensor
+            sensor_transform = carla.Transform(sensor_location, sensor_rotation)
+            sensor = CarlaDataProvider.get_world().spawn_actor(bp, sensor_transform, vehicle)
             # setup callback
             sensor.listen(CallBack(sensor_spec['id'], sensor, self._agent.sensor_interface))
             self._sensors_list.append(sensor)
