+++ srunner/autoagents/autonomous_agent.py
@@ -13,7 +13,7 @@ from enum import Enum
 
 import carla
 
-from srunner.challenge.envs.sensor_interface import SensorInterface
+from srunner.autoagents.sensor_interface import SensorInterface
 from srunner.scenariomanager.timer import GameTime
 from srunner.tools.route_manipulation import downsample_route
